lab_8/pyproject.toml
[project]
name = "lab_8"
version = "0.1.0"
description = "GLM-OCR Document Parsing (Ollama + Self-hosted SDK)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.1",
    "ollama",
    "requests",
    "pillow",
    "glmocr",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = []
